warning hunting for xcode 4.6

+++ src/BeanUpdater.h
@@ -87,7 +87,7 @@ HIBERLITE_STRINGIFY_UNSIGNED_NUM(unsigned long long int)
 
 
 		static std::string toSQLiteValue(const char* val){
-			int n=strlen(val);
+			int n=(int)strlen(val);
 			char *s = new char[4+2*n];
 			int x = sprintf(s,"X'");
